Restructuring GRASS-folium, improved docstrings, pylint, black
