+++ python/grass/jupyter/Makefile
@@ -11,7 +11,8 @@ MODULES = \
 	interact_display \
 	region \
 	render3d \
-	utils
+	utils \
+	reprojection_renderer
 
 PYFILES := $(patsubst %,$(DSTDIR)/%.py,$(MODULES) __init__)
 PYCFILES := $(patsubst %,$(DSTDIR)/%.pyc,$(MODULES) __init__)
